Reorg explain
